Cleanup Fix IBP version check in DelegationTokenControlManager
